+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## In development
+### Bugfixes
+- Terrain Ruler no longer blocks other tools ([#3](https://github.com/manuelVo/foundryvtt-terrain-ruler/issues/3))
+
 ## 1.0.0
 Initial release